add: initialization to separate file
